requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
mysql-connector-python
